Safari-609.1.11.1/LayoutTests/crypto/subtle/rsa-import-spki-key-export-spki-key-expected.txt

Test importing a SPKI RSA-OAEP key and then export it in SPKI format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS bytesToHexString(publicKey) is spkiKeyHex
PASS successfullyParsed is true

TEST COMPLETE
